pyproject.toml
[project]
name = "auto-coin-trader-v3"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "binance-sdk-derivatives-trading-usds-futures>=1.3.0",
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "slack-sdk>=3.36.0",
    "sqlmodel>=0.0.24",
    "streamlit>=1.49.1",
    "websockets>=15.0.1",
]
